fixed componenter & login page
