+++ src/components/Dashboard.tsx
@@ -2,11 +2,11 @@ import { useEffect, useState } from "react";
 import { getAppplicants } from "../services/fakeApplicantsService";
 import IApplicant from "../types/IApplicant";
 import Table from "../components/Table";
-import applicantsContext from "../context/applicantsContext";
-import Sidebar from "../Sidebar";
+import applicantsContext from "../Context/ApplicantsContext";
+import Sidebar from "../components/Sidebar";
 import SearchBar from "./SearchBar";
 
-function Items() {
+function Items(): JSX.Element {
   const [applicants, setApplicants] = useState<IApplicant[]>([]);
 
   useEffect(() => {
