+++ src/index.d.ts
@@ -1,1 +1,2 @@
 declare module "*.jpg";
+declare module "*.mp4";
